chore: changes gem structure through adding DomainDrivenToolkit::Patterns module
