+++ lib/domain_driven_toolkit.rb
@@ -2,11 +2,7 @@ require 'domain_driven_toolkit/version'
 
 module DomainDrivenToolkit
   require 'domain_driven_toolkit/application'
-  require 'domain_driven_toolkit/value'
-  require 'domain_driven_toolkit/entity'
-  require 'domain_driven_toolkit/service'
-  require 'domain_driven_toolkit/factory'
-  require 'domain_driven_toolkit/repository'
+  require 'domain_driven_toolkit/patterns'
 
   class << self
     def version
@@ -17,25 +13,8 @@ module DomainDrivenToolkit
       Application
     end
 
-    def value
-      Value
-    end
-
-    def entity
-      Entity
-    end
-
-    def service
-      Service
-    end
-
-    def factory
-      Factory
-    end
-
-    def repository
-      Repository
+    def patterns
+      Patterns
     end
   end
-
 end
