--- lib/domain_driven_toolkit/entity.rb
@@ -1,5 +1,0 @@
-module DomainDrivenToolkit
-  class Entity
-
-  end
-end
