--- lib/domain_driven_toolkit/factory.rb
@@ -1,5 +1,0 @@
-module DomainDrivenToolkit
-  class Factory
-
-  end
-end
