+++ lib/domain_driven_toolkit/patterns/entity.rb
@@ -1,0 +1,7 @@
+module DomainDrivenToolkit
+  module Patterns
+    class Entity
+
+    end
+  end
+end
