+++ lib/domain_driven_toolkit/patterns/factory.rb
@@ -1,0 +1,7 @@
+module DomainDrivenToolkit
+  module Patterns
+    class Factory
+
+    end
+  end
+end
