+++ lib/domain_driven_toolkit/patterns/repository.rb
@@ -1,0 +1,7 @@
+module DomainDrivenToolkit
+  module Patterns
+    class Repository
+
+    end
+  end
+end
