+++ lib/domain_driven_toolkit/patterns/service.rb
@@ -1,0 +1,7 @@
+module DomainDrivenToolkit
+  module Patterns
+    class Service
+
+    end
+  end
+end
