+++ lib/domain_driven_toolkit/patterns/value.rb
@@ -1,0 +1,7 @@
+module DomainDrivenToolkit
+  module Patterns
+    class Value
+
+    end
+  end
+end
