--- lib/domain_driven_toolkit/repository.rb
@@ -1,4 +1,0 @@
-module DomainDrivenToolkit
-  class Repository
-  end
-end
