--- lib/domain_driven_toolkit/service.rb
@@ -1,5 +1,0 @@
-module DomainDrivenToolkit
-  class Service
-
-  end
-end
