--- lib/domain_driven_toolkit/value.rb
@@ -1,5 +1,0 @@
-module DomainDrivenToolkit
-  class Value
-
-  end
-end
