--- spec/entity_spec.rb
@@ -1,3 +1,0 @@
-RSpec.describe DomainDrivenToolkit.entity do
-
-end
