--- spec/factory_spec.rb
@@ -1,3 +1,0 @@
-RSpec.describe DomainDrivenToolkit.factory do
-
-end
