+++ spec/patterns/entity_spec.rb
@@ -1,0 +1,3 @@
+RSpec.describe DomainDrivenToolkit.patterns.entity do
+
+end
