+++ spec/patterns/factory_spec.rb
@@ -1,0 +1,3 @@
+RSpec.describe DomainDrivenToolkit.patterns.factory do
+
+end
