+++ spec/patterns/repository_spec.rb
@@ -1,0 +1,3 @@
+RSpec.describe DomainDrivenToolkit.patterns.repository do
+
+end
