+++ spec/patterns/service_spec.rb
@@ -1,0 +1,3 @@
+RSpec.describe DomainDrivenToolkit.patterns.service do
+
+end
