+++ spec/patterns/value_spec.rb
@@ -1,0 +1,3 @@
+RSpec.describe DomainDrivenToolkit.patterns.value do
+
+end
