--- spec/service_spec.rb
@@ -1,3 +1,0 @@
-RSpec.describe DomainDrivenToolkit.service do
-
-end
